KVM: x86: Add AMD SEV specific Hypercall3
KVM hypercall framework relies on alternative framework to patch the
'VMCALL' to 'VMMCALL' on AMD platform. If a hypercall is made before
apply_alternative() is called then KVM hypercall framework defaults to
VMCALL. The approach works fine on non SEV guest because VMCALL would
causes #UD, and hypervisor is able to fetch the instruction by reading
the guest memory. But when SEV is active, guest memory is encrypted with
guest key and hypervisor will no longer able to access the guest memory
to decode the instruction bytes.

The patch adds SEV specific hypercall3, which unconditionally uses VMMCALL.
This hypercall will be used by SEV guest to notify which page is encrypted
in the guest memory. We need this information during guest migration or
debug time. e.g If page is encrypted then we need to use SEV firmware command
to re-encrypt using transport key during migration and similarly decrypt the
page during debug.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>