KVM: x86: Add AMD SEV specific Hypercall3
KVM hypercall framework relies on alternative framework to patch the
VMCALL -> VMMCALL on AMD platform. If a hypercall is made before
apply_alternative() is called then we default to VMCALL. The approach
works fine on non SEV guest. A VMCALL would causes #UD, and hypervisor
will be able to fetch the instruction by reading the guest memory to
decode the instruction. But when SEV is active, guest memory is encrypted
with guest key and hypervisor will not be able to decode the instruction
bytes.

Add SEV specific hypercall3, it unconditionally uses VMMCALL. The hypercall
will be used by the SEV guest to notify encrypted pages to the hypervisor.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>