KVM: X86: Introduce KVM_HC_SET_PAGE_ENC hypercall
The hypercall can be used by the SEV guest to notify the page encryption
attribute to the hypervisor. The hypercall should be invoked only when
the encryption attribute is changed from encrypted -> decrypted and vice
versa. By default all the guest pages should be considered encrypted.

The page encrypt bitmap can be referred during the SEV guest migration to
check whether the guest page contains encrypted or decrypted data. If page
is set as encrypted by the SEV guest then we must use the SEV specific
command to access the guest page during the migration.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>